refactor: Changed model IDs from int to long, we now pass a full station object through the addfillup views

+++ app/src/main/java/com/automotive/hhi/mileagetracker/model/data/CarFactory.java
@@ -16,7 +16,7 @@ public class CarFactory  {
                 .getDouble(cursor
                         .getColumnIndexOrThrow(DataContract.CarTable.AVGMPG)));
         car.setId(cursor
-                .getInt(cursor
+                .getLong(cursor
                         .getColumnIndexOrThrow(DataContract.CarTable._ID)));
         car.setMake(cursor
                 .getString(cursor
